requirements.txt
flask
flask-login 
flask-jwt-extended 
flask-bcrypt
flask_sqlalchemy
flask-socketio
SQLAlchemy
python-dotenv
psycopg2